move internal/computecore out of internal
Part of issue 2823
