+++ pkg/computecore/computecore.go
